requirements.txt
aiogram==3.13.1
aiohttp==3.9.5
deep_translator==1.11.4
Flask==3.0.3
httpx==0.27.2
mistralai==1.1.0
Pillow==10.4.0
psycopg2_binary==2.9.9
python-dotenv==1.0.1
Requests==2.32.3
